!whois xbox command added
